ml_services/requirements.txt
# PyTorch and ML frameworks
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.35.2
sentence-transformers==2.2.2
stable-baselines3==2.2.1

# Hugging Face and NLP
tokenizers==0.15.0
datasets==2.15.0
accelerate==0.25.0

# Vector search and similarity
faiss-cpu==1.7.4
annoy==1.17.3

# String matching and phonetics
jellyfish==1.0.1
phonetics==1.0.5
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0

# Data processing
numpy==1.26.2
pandas==2.1.3
polars==0.19.12
scikit-learn==1.3.2

# Task queue and async
celery==5.3.4
redis==5.0.1
asyncio==3.4.3

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
structlog==23.2.0
tqdm==4.66.1

# Monitoring
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1